Added distance-based mse loss function.
C++ code and python bindings for calculating the distance to the
closest point with a positive value in a ligand channel for each
spatial point in the input as well as updated build instructions which
now includes a CMakeLists.txt and more advanced setup.py.

Using --loss distance_mse for autoencoder/gnina_autoencoder.py will use
the distance-weighted mean squared error, with the distance defined as
above, the idea being that we care more about the reconstruction loss in
the region of protein-ligand contact than we do about protein
information far from the ligand.