Added distance-from-ligand-thresholded metrics.
The --metric_distance_threshold argument to gnina_autoencoder, if
specified, will cause the close_mae, close_nonzero_mae, and
close_zero_mae metrics to be calculated, reported and recorded.

These metrics are similar to mae, nonzero_mae and zero_mae in that they
provide the mean average error, and the MAE for zero and non-zero
inputs, but only inputs closer than metric_distance_threshold Angstroms
of any positive value for a ligand channel are considered.